backend/requirements.txt
python-dotenv==1.2.1
google-genai

Package              Version
-------------------- ----------
annotated-doc        0.0.4
annotated-types      0.7.0
anyio                4.12.0
cachetools           6.2.4
certifi              2025.11.12
charset-normalizer   3.4.4
click                8.3.1
colorama             0.4.6
distro               1.9.0
dnspython            2.8.0
email-validator      2.3.0
fastapi              0.128.0
fastapi-cli          0.0.20
fastapi-cloud-cli    0.8.0
fastar               0.8.0
google-auth          2.45.0
google-genai         1.56.0
h11                  0.16.0
httpcore             1.0.9
httptools            0.7.1
httpx                0.28.1
idna                 3.11
Jinja2               3.1.6
markdown-it-py       4.0.0
MarkupSafe           3.0.3
mdurl                0.1.2
pip                  25.2
pyasn1               0.6.1
pyasn1_modules       0.4.2
pydantic             2.12.5
pydantic_core        2.41.5
pydantic-extra-types 2.11.0
pydantic-settings    2.12.0
Pygments             2.19.2
python-dotenv        1.2.1
python-multipart     0.0.21
PyYAML               6.0.3
requests             2.32.5
rich                 14.2.0
rich-toolkit         0.17.1
rignore              0.7.6
rsa                  4.9.1
sentry-sdk           2.48.0
shellingham          1.5.4
sniffio              1.3.1
starlette            0.50.0
tenacity             9.1.2
typer                0.21.0
typing_extensions    4.15.0
typing-inspection    0.4.2
urllib3              2.6.2
uvicorn              0.40.0
watchfiles           1.1.1
websockets           15.0.1